+++ zebra-state/src/lib.rs
@@ -42,7 +42,7 @@ pub use config::{
 pub use constants::{state_database_format_version_in_code, MAX_BLOCK_REORG_HEIGHT};
 pub use error::{
     BoxError, CloneError, CommitBlockError, CommitSemanticallyVerifiedError,
-    DuplicateNullifierError, ValidateContextError,
+    DuplicateNullifierError, StateInitError, ValidateContextError,
 };
 pub use request::{
     CheckpointVerifiedBlock, CommitSemanticallyVerifiedBlockRequest, HashOrHeight, MappedRequest,