+++ zebra-state/src/service.rs
@@ -57,7 +57,7 @@ use crate::{
         watch_receiver::WatchReceiver,
     },
     BoxError, CheckpointVerifiedBlock, CommitSemanticallyVerifiedError, Config, KnownBlock,
-    ReadRequest, ReadResponse, Request, Response, SemanticallyVerifiedBlock,
+    ReadRequest, ReadResponse, Request, Response, SemanticallyVerifiedBlock, StateInitError,
 };
 
 pub mod block_iter;
@@ -324,7 +324,8 @@ impl StateService {
                     &network,
                     #[cfg(feature = "elasticsearch")]
                     true,
-                );
+                )
+                .expect("opening a read-write finalized state database should not fail");
                 timer.finish_desc("opening finalized state database");
 
                 let timer = CodeTimer::start();
@@ -1833,11 +1834,14 @@ pub async fn init(
 pub fn init_read_only(
     config: Config,
     network: &Network,
-) -> (
-    ReadStateService,
-    ZebraDb,
-    tokio::sync::watch::Sender<NonFinalizedState>,
-) {
+) -> Result<
+    (
+        ReadStateService,
+        ZebraDb,
+        tokio::sync::watch::Sender<NonFinalizedState>,
+    ),
+    StateInitError,
+> {
     let finalized_state = FinalizedState::new_with_debug(
         &config,
         network,
@@ -1845,11 +1849,11 @@ pub fn init_read_only(
         #[cfg(feature = "elasticsearch")]
         false,
         true,
-    );
+    )?;
     let (non_finalized_state_sender, non_finalized_state_receiver) =
         tokio::sync::watch::channel(NonFinalizedState::new(network));
 
-    (
+    Ok((
         ReadStateService::new(
             &finalized_state,
             None,
@@ -1857,7 +1861,7 @@ pub fn init_read_only(
         ),
         finalized_state.db.clone(),
         non_finalized_state_sender,
-    )
+    ))
 }
 
 /// Calls [`init_read_only`] with the provided [`Config`] and [`Network`] from a blocking task.
@@ -1865,11 +1869,16 @@ pub fn init_read_only(
 pub fn spawn_init_read_only(
     config: Config,
     network: &Network,
-) -> tokio::task::JoinHandle<(
-    ReadStateService,
-    ZebraDb,
-    tokio::sync::watch::Sender<NonFinalizedState>,
-)> {
+) -> tokio::task::JoinHandle<
+    Result<
+        (
+            ReadStateService,
+            ZebraDb,
+            tokio::sync::watch::Sender<NonFinalizedState>,
+        ),
+        StateInitError,
+    >,
+> {
     let network = network.clone();
     tokio::task::spawn_blocking(move || init_read_only(config, &network))
 }